LayoutTests/plugins/reloadplugins-and-pages-expected.txt

ALERT: Page was loaded
ALERT: Page was loaded

Test for NPN_RelaodPlugin in which reloadOpenPages is true. You should see two alerts, indicating that the page was loaded twice.

SUCCESS
